check if OpenInTerminal-Lite is already installed

+++ macsetup.sh
@@ -1311,15 +1311,17 @@ fi
 #
 # https://github.com/Ji4n1ng/OpenInTerminal
 #
-installcask openinterminal-lite
-defaults write wang.jianing.app.OpenInTerminal-Lite LiteDefaultTerminal iTerm
-open /Applications
-notification "drag openinterminal-lite in Finder toolbar while pressing Command"
-curl -L https://github.com/Ji4n1ng/OpenInTerminal/releases/download/v1.2.0/Icons.zip  --output icons.zip
-unzip icons.zip
-rm icons.zip
-seticon icons/icon_terminal_dark.icns /Applications/OpenInTerminal-Lite.app
-rm -rf icons
+if [[ ! $(isAppInstalled OpenInTerminal-Lite) ]]; then
+    installcask openinterminal-lite
+    defaults write wang.jianing.app.OpenInTerminal-Lite LiteDefaultTerminal iTerm
+    open /Applications
+    notification "drag openinterminal-lite in Finder toolbar while pressing Command"
+    curl -L https://github.com/Ji4n1ng/OpenInTerminal/releases/download/v1.2.0/Icons.zip  --output icons.zip
+    unzip icons.zip
+    rm icons.zip
+    seticon icons/icon_terminal_dark.icns /Applications/OpenInTerminal-Lite.app
+    rm -rf icons
+fi
 
 #
 # Rain
